+++ src/Drivers/FileDriver.php
@@ -12,12 +12,14 @@ class FileDriver extends Driver
 {
     protected array $settings;
     protected array $options;
+    protected array $modelConfig;
     protected array $tags;
 
-    public function __construct(array $settings, array $options, array $tags = [])
+    public function __construct(array $settings, array $options, array $modelConfig, array $tags = [])
     {
         $this->settings = $settings;
         $this->options = $options;
+        $this->modelConfig = $modelConfig;
         $this->tags = $tags;
     }